+++ src/main/java/com/yuuko/core/commands/Command.java
@@ -1,6 +1,10 @@
 package com.yuuko.core.commands;
 
 import com.yuuko.core.Configuration;
+import com.yuuko.core.database.DatabaseFunctions;
+import com.yuuko.core.utilities.MessageHandler;
+import com.yuuko.core.utilities.Utils;
+import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 import org.slf4j.Logger;
@@ -51,6 +55,28 @@ public abstract class Command {
 
     public Permission[] getPermissions() {
         return permissions;
+    }
+
+    public boolean checkCommandSettings(MessageReceivedEvent e) {
+        // Executor still checks core/developer, in this case simply return true.
+        String moduleName = Utils.getModuleName(module);
+        if(moduleName.equals("Core") || moduleName.equals("Developer")) {
+            return true;
+        }
+
+        // Checks if the command is globally disabled either global:global, global:local, local:global or local:local.
+        String guild = e.getGuild().getId();
+        String channel = e.getTextChannel().getId();
+        if(!DatabaseFunctions.checkCommandSettings("*", guild, "*") &&
+           !DatabaseFunctions.checkCommandSettings("*", guild, channel) &&
+           !DatabaseFunctions.checkCommandSettings(name, guild, "*") &&
+           !DatabaseFunctions.checkCommandSettings(name, guild, channel)) {
+            return true;
+        } else {
+            EmbedBuilder embed = new EmbedBuilder().setTitle("Command Disabled").setDescription("The `" + name + "` command is disabled.");
+            MessageHandler.sendMessage(e, embed.build());
+            return false;
+        }
     }
 
     // Abstract method signature to ensure method is implemented.