+++ src/main/java/com/yuuko/core/commands/world/commands/WeatherCommand.java
@@ -10,6 +10,7 @@ import com.yuuko.core.utilities.json.JsonBuffer;
 import net.dv8tion.jda.core.EmbedBuilder;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 
+import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
@@ -47,6 +48,7 @@ public class WeatherCommand extends Command {
                     .addField("Sunset", LocalDateTime.ofEpochSecond(data.get("sys").getAsJsonObject().get("sunset").getAsInt(), 0, ZoneOffset.UTC).format(DateTimeFormatter.ofPattern("hh:mma")), true)
                     .addField("Longitude", data.get("coord").getAsJsonObject().get("lon").getAsString(), true)
                     .addField("Wind Angle", data.get("wind").getAsJsonObject().get("deg").getAsString() + "°", true)
+                    .setTimestamp(Instant.now())
                     .setFooter(Configuration.STANDARD_STRINGS[1] + e.getMember().getEffectiveName(), e.getAuthor().getEffectiveAvatarUrl());
             MessageHandler.sendMessage(e, embed.build());